remove double CUBE

+++ tools_casting.c
@@ -6,7 +6,7 @@
 /*   By: oufisaou <oufisaou@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/20 16:14:40 by oufisaou          #+#    #+#             */
-/*   Updated: 2023/02/14 13:56:47 by oufisaou         ###   ########.fr       */
+/*   Updated: 2023/02/14 14:03:00 by oufisaou         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 void	begining_horizontal(t_all *cub)
 {
 	what_direction(cub);
-	cub->var_d.yinter = floor(cub->player.y / CUBE) * (double)CUBE;
+	cub->var_d.yinter = floor(cub->player.y / CUBE) * CUBE;
 	is_down(cub);
 	cub->var_d.xinter = cub->player.x + \
 	((cub->var_d.yinter - cub->player.y) / tan(cub->var_d.new_angle));
